refactor: added SubcriptionManager
