+++ apps/relay/src/server/track.rs
@@ -18,26 +18,16 @@ use crate::server::config::AppConfig;
 use crate::server::object_logger::ObjectLogger;
 use crate::server::stream_id::StreamId;
 use crate::server::subscription::Subscription;
+use crate::server::subscription_manager::SubscriptionManager;
 use crate::server::utils;
 use anyhow::Result;
 use moqtail::model::common::location::Location;
 use moqtail::model::control::subscribe::Subscribe;
 use moqtail::model::data::object::Object;
 use moqtail::{model::common::tuple::Tuple, transport::data_stream_handler::HeaderInfo};
-use std::{collections::BTreeMap, collections::HashMap, sync::Arc};
+use std::sync::Arc;
 use tokio::sync::RwLock;
-use tokio::sync::mpsc::UnboundedSender;
 use tracing::{debug, error, info};
-
-/// Number of partitions for subscriber senders management to reduce lock contention.
-/// Each partition contains a separate HashMap protected by its own RwLock.
-/// Higher values reduce contention but increase memory overhead.
-/// Should be a power of 2 for optimal modulo performance.
-const SUBSCRIBER_PARTITION_COUNT: usize = 16;
-
-pub type SubscriberSenderMap = HashMap<usize, UnboundedSender<TrackEvent>>;
-pub type SubscriberSenderLock = Arc<RwLock<SubscriberSenderMap>>;
-pub type SubscriberSenderList = Vec<SubscriberSenderLock>;
 
 #[derive(Debug, Clone)]
 #[allow(clippy::large_enum_variant)]
@@ -62,14 +52,12 @@ pub struct Track {
   pub track_namespace: Tuple,
   #[allow(dead_code)]
   pub track_name: String,
-  subscriptions: Arc<RwLock<BTreeMap<usize, Arc<RwLock<Subscription>>>>>,
+  pub subscription_manager: SubscriptionManager,
   pub publisher_connection_id: usize,
   #[allow(dead_code)]
   pub(crate) cache: TrackCache,
-  subscriber_senders: Arc<SubscriberSenderList>,
   pub largest_location: Arc<RwLock<Location>>,
   pub object_logger: ObjectLogger,
-  log_folder: String,
   config: &'static AppConfig,
 }
 
@@ -83,44 +71,21 @@ impl Track {
     publisher_connection_id: usize,
     config: &'static AppConfig,
   ) -> Self {
-    // Initialize partitioned subscriber senders
-    let mut subscriber_senders = Vec::with_capacity(SUBSCRIBER_PARTITION_COUNT);
-    for _ in 0..SUBSCRIBER_PARTITION_COUNT {
-      subscriber_senders.push(Arc::new(RwLock::new(HashMap::new())));
-    }
-
     Track {
       track_alias,
       track_namespace,
       track_name,
-      subscriptions: Arc::new(RwLock::new(BTreeMap::new())),
+      subscription_manager: SubscriptionManager::new(
+        track_alias,
+        config.log_folder.clone(),
+        config,
+      ),
       publisher_connection_id,
       cache: TrackCache::new(track_alias, config.cache_size.into(), config),
-      subscriber_senders: Arc::from(subscriber_senders),
       largest_location: Arc::new(RwLock::new(Location::new(0, 0))),
       object_logger: ObjectLogger::new(config.log_folder.clone()),
-      log_folder: config.log_folder.clone(),
       config,
     }
-  }
-
-  /// Calculate the partition index for subscriber distribution across buckets.
-  /// This method implements a load balancing strategy to distribute subscribers
-  /// across multiple buckets to improve performance and reduce contention.
-  fn get_subscriber_partition_index(&self, subscriber_id: usize) -> usize {
-    // Convert to bytes for fnv_hash function
-    let value_bytes = subscriber_id.to_le_bytes();
-    (utils::fnv_hash(&value_bytes) % SUBSCRIBER_PARTITION_COUNT as u64) as usize
-  }
-
-  /// Get the subscriber sender for a specific subscriber
-  async fn _get_subscriber_sender(
-    &self,
-    subscriber_id: usize,
-  ) -> Option<UnboundedSender<TrackEvent>> {
-    let partition_index = self.get_subscriber_partition_index(subscriber_id);
-    let senders = self.subscriber_senders[partition_index].read().await;
-    senders.get(&subscriber_id).cloned()
   }
 
   pub async fn add_subscription(
@@ -128,83 +93,26 @@ impl Track {
     subscriber: Arc<MOQTClient>,
     subscribe_message: Subscribe,
   ) -> Result<(), anyhow::Error> {
-    let connection_id = { subscriber.connection_id };
-
-    info!(
-      "Adding subscription for subscriber_id: {} to track: {} subscription message: {:?}",
-      connection_id, self.track_alias, subscribe_message
-    );
-
-    // Create a separate unbounded channel for this subscriber
-    let (event_tx, event_rx) = tokio::sync::mpsc::unbounded_channel::<TrackEvent>();
-
-    let subscription = Subscription::new(
-      self.track_alias,
-      subscribe_message,
-      subscriber.clone(),
-      event_rx,
-      self.cache.clone(),
-      connection_id,
-      self.log_folder.clone(),
-      self.config,
-    );
-
-    let mut subscriptions = self.subscriptions.write().await;
-    if subscriptions.contains_key(&connection_id) {
-      error!(
-        "Subscriber with connection_id: {} already exists in track: {}",
-        connection_id, self.track_alias
-      );
-      return Err(anyhow::anyhow!("Subscriber already exists"));
-    }
-    subscriptions.insert(connection_id, Arc::new(RwLock::new(subscription)));
-
-    // Store the sender for this subscriber in the appropriate partition
-    let partition_index = self.get_subscriber_partition_index(connection_id);
-    let mut senders = self.subscriber_senders[partition_index].write().await;
-    senders.insert(connection_id, event_tx);
-
-    Ok(())
+    self
+      .subscription_manager
+      .add_subscription(subscriber, subscribe_message, self.cache.clone())
+      .await
   }
 
   // return the subscription for the client
   // subscriber_id is the connection id of the client
   pub async fn get_subscription(&self, subscriber_id: usize) -> Option<Arc<RwLock<Subscription>>> {
-    debug!(
-      "Getting subscription for subscriber_id: {} from track: {}",
-      subscriber_id, self.track_alias
-    );
-    let subscriptions = self.subscriptions.read().await;
-    // find the subscription by subscriber_id and finish it
-    subscriptions.get(&subscriber_id).cloned()
+    self
+      .subscription_manager
+      .get_subscription(subscriber_id)
+      .await
   }
 
   pub async fn remove_subscription(&self, subscriber_id: usize) {
-    info!(
-      "Removing subscription for subscriber_id: {} from track: {}",
-      subscriber_id, self.track_alias
-    );
-    let mut subscriptions = self.subscriptions.write().await;
-    let sub = subscriptions.remove(&subscriber_id);
-    drop(subscriptions);
-
-    // find the subscription by subscriber_id and finish it
-    if let Some(subscription) = sub {
-      let sub = subscription.write().await;
-      sub.finish().await;
-    }
-
-    // Remove and dispose the sender for this subscriber from the appropriate partition
-    let partition_index = self.get_subscriber_partition_index(subscriber_id);
-    let mut senders = self.subscriber_senders[partition_index].write().await;
-    if let Some(_sender) = senders.remove(&subscriber_id) {
-      // The sender is automatically dropped here, which closes the channel
-      info!(
-        "Disposed sender for subscriber_id: {} from track: {}",
-        subscriber_id, self.track_alias
-      );
-    }
-    drop(senders);
+    self
+      .subscription_manager
+      .remove_subscription(subscriber_id)
+      .await
   }
 
   pub async fn new_object(
@@ -262,7 +170,10 @@ impl Track {
         header_info: header_info.cloned(),
       };
 
-      self.send_event_to_subscribers(event).await?;
+      self
+        .subscription_manager
+        .send_event_to_subscribers(event)
+        .await?;
       Ok(())
     } else {
       error!(
@@ -282,7 +193,10 @@ impl Track {
       stream_id: stream_id.clone(),
     };
 
-    self.send_event_to_subscribers(event).await?;
+    self
+      .subscription_manager
+      .send_event_to_subscribers(event)
+      .await?;
 
     Ok(())
   }
@@ -298,54 +212,12 @@ impl Track {
       reason: "Publisher disconnected".to_string(),
     };
 
-    self.send_event_to_subscribers(event).await?;
+    self
+      .subscription_manager
+      .send_event_to_subscribers(event)
+      .await?;
 
     Ok(())
-  }
-
-  // Send event to all subscribers
-  async fn send_event_to_subscribers(
-    &self,
-    event: TrackEvent,
-  ) -> Result<Vec<usize>, anyhow::Error> {
-    let mut failed_subscribers = Vec::new();
-    let mut total_subscribers = 0;
-
-    // Iterate through all partitions
-    for partition in self.subscriber_senders.iter() {
-      let senders = partition.read().await;
-
-      if !senders.is_empty() {
-        total_subscribers += senders.len();
-
-        for (subscriber_id, sender) in senders.iter() {
-          if let Err(e) = sender.send(event.clone()) {
-            error!(
-              "Failed to send event to subscriber {}: {}",
-              subscriber_id, e
-            );
-            failed_subscribers.push(*subscriber_id);
-          }
-        }
-      }
-    }
-
-    if !failed_subscribers.is_empty() {
-      error!(
-        "{:?} event sent to {} subscribers, {} failed for track: {}",
-        event,
-        total_subscribers - failed_subscribers.len(),
-        failed_subscribers.len(),
-        self.track_alias
-      );
-    } else if total_subscribers > 0 {
-      debug!(
-        "{:?} event sent successfully to {} subscribers for track: {}",
-        event, total_subscribers, self.track_alias
-      );
-    }
-
-    Ok(failed_subscribers)
   }
 }
